Meilleur support des bases sur FS
BUGS :
    - RAC 12.2  : impossible de créér une base.
    - STBY 12.2 : le duplicate plante sur
        RMAN-00571: ===========================================================
        RMAN-00569: =============== ERROR MESSAGE STACK FOLLOWS ===============
        RMAN-00571: ===========================================================
        RMAN-03002: failure of Duplicate Db command at 03/13/2017 17:29:16
        RMAN-05501: aborting duplication of target database
        RMAN-03015: error occurred in stored script Memory Script
        RMAN-03009: failure of sql command on clone_default channel at 03/13/2017 17:29:16
        RMAN-11003: failure during parse/execution of SQL statement: alter database mount standby database
        ORA-01103: le nom de base de donnees 'PLUTOT' dans le fichier de controle n'est pas 'PLUTARD'

Fichiers modifiées :
    database_servers/clone_master.sh
    database_servers/create_systemd_service_oracledb.sh
    database_servers/define_new_server.sh
    database_servers/install_oracle.sh
    database_servers/oracle_uninstall.sh
    database_servers/show_info_server.sh
    database_servers/uninstall.sh
    db/configure_backup.sh
    db/create_db.sh
    db/create_pdb.sh
    db/create_srv_for_single_db.sh
    db/drop_db.sh
    db/fsdb_manager.sh
    db/remove_all_files_for_db.sh
    db/rman/set_config.rman
    db/sql/dbv.sql
    dblib.sh
    dg/drop_afd_disks.sh
    disk/create_fs.sh
    disk/iostat_on_bdd_disks.sh
    global.cfg
    plelib.sh
    db/sql/create_trigger_start_pdb_services.sql

+++ database_servers/oracle_uninstall.sh
@@ -49,13 +49,16 @@ EOS
 fi
 LN
 
-# En 12.2 l'ORACLE_HOME des autres nœuds n'est pas purgé.
-while read node
-do
-	[[ x"$node" == x || "$node" == $(hostname -s) ]] && continue || true
+if test_if_cmd_exists olsnodes
+then
+	# En 12.2 l'ORACLE_HOME des autres nœuds ne sont pas supprimés.
+	while read node
+	do
+		[[ x"$node" == x || "$node" == $(hostname -s) ]] && continue || true
 
-	exec_cmd "ssh $node '. .bash_profile && [ -d \$ORACLE_HOME ] && rm -rf \$ORACLE_HOME/* || true'<</dev/null"
-	LN
-done<<<"$(olsnodes)"
+		exec_cmd "ssh $node '. .bash_profile && [ -d \$ORACLE_HOME ] && rm -rf \$ORACLE_HOME/* || true'<</dev/null"
+		LN
+	done<<<"$(olsnodes)"
+fi
 
 exit 0
